[#51] 재고 복구 이벤트 리스너 구현

+++ reward/src/main/java/com/nowayback/reward/domain/reward/command/CreateRewardOptionCommand.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public record CreateRewardOptionCommand(
         String name,
-        Integer additionalPrice,
+        Long additionalPrice,
         Integer stockQuantity,
         Boolean isRequired,
         Integer displayOrder
